.legacy/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
bitsandbytes==0.46.1
certifi==2025.7.14
charset-normalizer==3.4.2
comm==0.2.3
datasets==4.0.0
debugpy==1.8.15
decorator==5.2.1
diffusers==0.20.0
dill==0.3.8
einops==0.8.1
executing==2.2.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
setuptools==70.2.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
transformers-stream-generator==0.0.5
triton==3.3.1
trl==0.19.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
